Small reorganization of files, code and folders.
