+++ src/pages/loginsignup/LoginSignup.css
@@ -2,10 +2,7 @@
 body{
     display: flex;
     justify-content: center;
-    background: linear-gradient(#1584e4, #e83186);
-    height: 100%;
-      color: #000000;
-
+  
 }
 
 .icon {
